# docs/openapi-snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jekko-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`GET /api/v1/session?project_id=…` — list sessions for a project.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Sessions"
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`POST /api/v1/session` — create a session.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created session"
          }
        }
      }
    },
    "/api/v1/session/{id}": {
      "get": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`GET /api/v1/session/:id`.",
        "operationId": "get_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`DELETE /api/v1/session/:id` — best-effort delete (no-op in-memory).",
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/v1/session/{id}/abort": {
      "post": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`POST /api/v1/session/:id/abort` — request that a running session abort.",
        "operationId": "abort",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acknowledged"
          }
        }
      }
    },
    "/api/v1/session/{id}/message": {
      "get": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`GET /api/v1/session/:id/message`.",
        "operationId": "messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages"
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::session"
        ],
        "summary": "`POST /api/v1/session/:id/message`.",
        "operationId": "append_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendMessageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Appended"
          }
        }
      }
    },
...
    "/api/v1/sync/replay": {
      "post": {
        "tags": [
          "crate::routes::sync"
        ],
        "summary": "`POST /api/v1/sync/replay` — accept a batch of events to replay.",
        "operationId": "replay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {}
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replayed"
          }
        }
      }
    },
...
    "/api/v2/session/{id}/compact": {
      "post": {
        "tags": [
          "crate::routes::v2::session"
        ],
        "summary": "`POST /api/v2/session/:id/compact`.",
        "operationId": "compact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compacted"
          }
        }
      }
    },
    "/api/v2/session/{id}/message": {
      "get": {
        "tags": [
          "crate::routes::v2::message"
        ],
        "summary": "`GET /api/v2/session/:id/message`.",
        "operationId": "list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message page"
          }
        }
      }
    },
    "/api/v2/session/{id}/prompt": {
      "post": {
        "tags": [
          "crate::routes::v2::session"
        ],
        "summary": "`POST /api/v2/session/:id/prompt`.",
        "operationId": "prompt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/v2/session/{id}/wait": {
      "post": {
        "tags": [
          "crate::routes::v2::session"
        ],
        "summary": "`POST /api/v2/session/:id/wait`.",
        "operationId": "wait",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waited"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AppendMessageBody": {
        "type": "object",
        "description": "Request body for `POST /api/v1/session/:id/message`.",
        "required": [
          "role",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Free-form payload (JSON or string)."
          },
          "role": {
            "type": "string",
            "description": "Role."
          }
        }
      },
...
      "CreateSessionBody": {
        "type": "object",
        "description": "Request body for `POST /api/v1/session`.",
        "required": [
          "project_id",
          "directory"
        ],
        "properties": {
          "directory": {
            "type": "string",
            "description": "Working directory."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Parent session id."
          },
          "project_id": {
            "type": "string",
            "description": "Project id."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Title override."
          },
          "workspace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Workspace id."
          }
        }
      },
...
      "PromptBody": {
        "type": "object",
        "description": "Body of `POST /api/v2/session/:id/prompt`.",
        "required": [
          "prompt"
        ],
        "properties": {
          "delivery": {
            "type": [
              "string",
              "null"
            ],
            "description": "Delivery mode (`immediate` / `deferred`)."
          },
          "prompt": {
            "type": "object",
            "description": "User prompt (jekko-core `Prompt` shape, serialised as JSON)."
          }
        }
      },
...
    }
  }
}